@@ -1,24 +1,5 @@
+import { createRoot } from 'react-dom/client'
+import App from './App.tsx'
+import './index.css'
 
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
-import App from './app/layout'
-import HomePage from './app/page'
-import AppointmentsPage from './app/appointments/page'
-import MedicationsPage from './app/medications/page'
-import './app/globals.css'
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route index element={<HomePage />} />
-          <Route path="appointments" element={<AppointmentsPage />} />
-          <Route path="medications" element={<MedicationsPage />} />
-          <Route path="*" element={<Navigate to="/" replace />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
-  </React.StrictMode>,
-)
+createRoot(document.getElementById("root")!).render(<App />);
