@@ -1,15 +1,22 @@
-
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import { componentTagger } from 'lovable-tagger'
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react-swc";
+import path from "path";
+import { componentTagger } from "lovable-tagger";
 
 // https://vitejs.dev/config/
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
+  server: {
+    host: "::",
+    port: 8080,
+  },
   plugins: [
     react(),
-    componentTagger()
-  ],
-  server: {
-    port: 8080
-  }
-})
+    mode === 'development' &&
+    componentTagger(),
+  ].filter(Boolean),
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
+}));
